First Trace Target added
In Preparation of Beam spread over enemies, GC_ShockLoop now apply Gameplay Cue to actor under hit mouse. We now can use the info of knowing if the Mouse Hit is on an Actor or the environment

PS : For testing purposes, Aura has the Electrocute GA as a start ability bound to RMD input. It'll be revoked in the future